ux: place typeclass instances in companion objects
