+++ examples/catsValidation/src/io/github/iltotore/iron/catsValidation/Main.scala
@@ -3,12 +3,10 @@ package io.github.iltotore.iron.catsValidation
 import cats.data.EitherNel
 import cats.effect.{IO, IOApp}
 import cats.syntax.all.*
-import io.github.iltotore.iron.{given, *}
+import io.github.iltotore.iron.*
 import io.github.iltotore.iron.cats.*
-import io.github.iltotore.iron.constraint.numeric.given
+import io.github.iltotore.iron.constraint.collection.Contain
 import io.github.iltotore.iron.constraint.numeric.Greater
-import io.github.iltotore.iron.constraint.string.given
-import io.github.iltotore.iron.constraint.string.Contain
 
 object Main extends IOApp.Simple:
