+++ examples/formZio/src/io/github/iltotore/iron/formZio/Account.scala
@@ -1,4 +1,6 @@
-import io.github.iltotore.iron.{*, given}, constraint.numeric.{*, given}, constraint.string.{*, given}
+import io.github.iltotore.iron.{*, given}
+import io.github.iltotore.iron.constraint.all.*
+
 import io.github.iltotore.iron.zioJson.given
 
 import zio.json.*
@@ -12,17 +14,17 @@ type Password = (Match["[A-Za-z].*[0-9]|[0-9].*[A-Za-z]"] & MinLength[6] & MaxLe
 type Age = Greater[0] DescribedAs "Age should be strictly positive"
 
 /**
-  * A basic Account with a name, a password and an age.
-  *
-  * @param name an alphanumeric String with a length between 3 and 10.
-  * @param password an String containing atleast a letter, a digit, with a length between 6 and 20.
-  * @param age a strictly positive Int.
-  */
+ * A basic Account with a name, a password and an age.
+ *
+ * @param name an alphanumeric String with a length between 3 and 10.
+ * @param password an String containing atleast a letter, a digit, with a length between 6 and 20.
+ * @param age a strictly positive Int.
+ */
 case class Account(name: String :| Username, password: String :| Password, age: Int :| Age)
 
 object Account:
 
   /**
-    * Account JSON encoder and decoder.
-    */
+   * Account JSON encoder and decoder.
+   */
   given JsonCodec[Account] = DeriveJsonCodec.gen
