+++ main/src/io/github/iltotore/iron/constraint/all.scala
@@ -1,0 +1,6 @@
+package io.github.iltotore.iron.constraint
+
+object all:
+  export any.*
+  export numeric.*
+  export string.*
