+++ main/test/src/io/github/iltotore/iron/testing/NumericSuite.scala
@@ -1,8 +1,15 @@
 package io.github.iltotore.iron.testing
-import io.github.iltotore.iron.*, constraint.numeric.{*, given}
+
+import io.github.iltotore.iron.{*, given}
+import io.github.iltotore.iron.constraint.numeric.*
 import utest.*
 
 object NumericSuite extends TestSuite:
+
+  type GtEqual[V] = (Greater[V] | StrictEqual[V]) DescribedAs ("Should be greater than or equal to V")
+
+  summon[Constraint[Int, Greater[0]]]
+  summon[Constraint[Int, GtEqual[0]]]
 
   val tests: Tests = Tests {
